Handle files that are deleted and recreated
If vim is set up to use a swap file, then when a save is changed, it
deletes the original file and renames the swap file into place.  This
can sometimes cause chokidar to send two events: "remove" and "add".
Because the DirectoryWatcher doesn't clear out removed files, the
subsequent add event is ignored and the change is not seen.