Raise deep directory changes after initial scan
For nested watchers, if the parent watcher or nested watcher has
finished its initial scan emit any new changes on the directory's
watcher. This is similar to each shallow directory watcher which emits
a change event for every file change after the initial scan of that
directory. The nested watchers though need to consider if the parent
finished its initial scan if so, any files found during the initial
scan for the nested directory should also be emitted as this is a new
directory created after the parent's initial scan.